third_party/WebKit/LayoutTests/webaudio/delaynode-maxdelaylimit-expected.txt

Tests attribute and maximum allowed delay of DelayNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.createDelay(180) threw exception NotSupportedError: Failed to execute 'createDelay' on 'AudioContext': The max delay time provided (180) is outside the range (0, 180)..
PASS context.createDelay(0) threw exception NotSupportedError: Failed to execute 'createDelay' on 'AudioContext': The max delay time provided (0) is outside the range (0, 180)..
PASS context.createDelay(-1) threw exception NotSupportedError: Failed to execute 'createDelay' on 'AudioContext': The max delay time provided (-1) is outside the range (0, 180)..
PASS context.createDelay(NaN) threw exception TypeError: Failed to execute 'createDelay' on 'AudioContext': The provided double value is non-finite..
PASS delay.delayTime.value === 0.5 is true
PASS Test signal was correctly delayed by 0.5 sec.
PASS successfullyParsed is true

TEST COMPLETE
